+++ src/app/api/__tests__/api-encounters.test.ts
@@ -1,7 +1,7 @@
 import { NextRequest } from "next/server";
 import { describe, expect, it } from "vitest";
 import { GET } from "@/app/api/encounters/route";
-import { RouteEncountersArraySchema } from "@/loaders/encounters";
+import { RouteEncountersArraySchema } from "@/types/encounters";
 
 describe("Encounters API", () => {
   it("should include egg locations with -1 ID", async () => {
