+++ src/components/PokemonCombobox/__tests__/encounterSelection.test.ts
@@ -1,6 +1,6 @@
 import { describe, expect, it } from "vitest";
-import { EncounterSource } from "@/loaders/encounters";
 import { type PokemonOptionType, PokemonStatus } from "@/loaders/pokemon";
+import { EncounterSource } from "@/types/encounters";
 import {
   applyEncounterDefaultStatus,
   getPokemonSources,
